Extract codex-instructions crate
Move contextual user fragment helpers and user instruction payload types into a dedicated crate, and rewire codex-core to consume it.
