+++ codex-rs/instructions/src/user_instructions_tests.rs
@@ -1,6 +1,10 @@
 use super::*;
 use codex_protocol::models::ContentItem;
+use codex_protocol::models::ResponseItem;
 use pretty_assertions::assert_eq;
+
+use crate::fragment::AGENTS_MD_FRAGMENT;
+use crate::fragment::SKILL_FRAGMENT;
 
 #[test]
 fn test_user_instructions() {
